+++ assets/javascripts/kitten/components/structure/info-lines/stories.js
@@ -82,14 +82,14 @@ WithFormComponents.args = {
   ...args,
   infos: [
     {
-      key: <Text size="tiny">Lorem importString docs size amet</Text>,
+      key: <Text size="small">Lorem importString docs size amet</Text>,
       value: (
         <TextInput name="input-item-1" size="tiny" placeholder="value 1" />
       ),
       id: 'item-1',
     },
     {
-      key: <Text size="tiny">Dolor</Text>,
+      key: <Text size="small">Dolor</Text>,
       value: (
         <TextInput name="input-item-2" size="tiny" placeholder="value 2" />
       ),