+++ assets/javascripts/kitten/karl/pages/subscription-details/stories.js
@@ -27,18 +27,18 @@ const PaymentList = ({
   <div className="k-SubscriptionDetails__listWrapper" {...props}>
     <Text
       weight={subtitleWeight}
-      size="tiny"
+      size="small"
       className="k-SubscriptionDetails__subtitle"
     >
       {subtitle}
     </Text>
-    <Text size="tiny" className="k-SubscriptionDetails__amount">
+    <Text size="small" className="k-SubscriptionDetails__amount">
       {amount}
     </Text>
     <Tag type={typeTag} size="small" className="k-SubscriptionDetails__tag">
       {tag}
     </Tag>
-    <Text size="tiny" className="k-SubscriptionDetails__date">
+    <Text size="small" className="k-SubscriptionDetails__date">
       {date}
     </Text>
   </div>
@@ -119,13 +119,13 @@ export const ContributorSubscription = () => {
           <img src={`/kitten-${Math.floor(Math.random() * 10)}.jpg`} alt="" />
         </div>
         <div>
-          <Text weight="regular" size="tiny">
+          <Text weight="regular" size="small">
             Club des lecteurs
           </Text>
-          <Text size="tiny" weight="light">
+          <Text size="small" weight="light">
             <br />
             par{' '}
-            <Text weight="bold" size="regular">
+            <Text weight="bold" size="medium">
               Sick Magazine
             </Text>
           </Text>
@@ -137,10 +137,10 @@ export const ContributorSubscription = () => {
               />
             }
           >
-            <Text size="tiny">Le nouveau numéro</Text>
-            <Text size="tiny">Les coulisses de l'édition</Text>
-            <Text size="tiny">Un poster édition limitée</Text>
-            <Text size="tiny">Une surprise collector</Text>
+            <Text size="small">Le nouveau numéro</Text>
+            <Text size="small">Les coulisses de l'édition</Text>
+            <Text size="small">Un poster édition limitée</Text>
+            <Text size="small">Une surprise collector</Text>
           </List>
         </div>
       </div>
@@ -148,14 +148,14 @@ export const ContributorSubscription = () => {
       <Separator className="k-u-margin-top-singleHalf k-u-margin-bottom-singleHalf" />
       <div className="k-SubscriptionDetails__details">
         <div>
-          <Text size="tiny">Fréquence</Text>
-          <Text tag="p" weight="bold" size="tiny" className="k-u-margin-none">
+          <Text size="small">Fréquence</Text>
+          <Text tag="p" weight="bold" size="small" className="k-u-margin-none">
             Mensuel
           </Text>
         </div>
         <div>
-          <Text size="tiny">Mode de livraison</Text>
-          <Text tag="p" weight="bold" size="tiny" className="k-u-margin-none">
+          <Text size="small">Mode de livraison</Text>
+          <Text tag="p" weight="bold" size="small" className="k-u-margin-none">
             Livraison
           </Text>
         </div>
@@ -163,8 +163,8 @@ export const ContributorSubscription = () => {
 
       <Separator className="k-u-margin-top-singleHalf k-u-margin-bottom-singleHalf" />
       <div className="k-SubscriptionDetails__option">
-        <Text size="tiny">Option</Text>
-        <Text tag="p" weight="bold" size="tiny" className="k-u-margin-none">
+        <Text size="small">Option</Text>
+        <Text tag="p" weight="bold" size="small" className="k-u-margin-none">
           Végétarien
         </Text>
       </div>
@@ -185,33 +185,33 @@ export const ContributorReward = () => {
               />
             </div>
             <div>
-              <Text weight="regular" size="tiny">
+              <Text weight="regular" size="small">
                 Club des lecteurs
               </Text>
-              <Text size="tiny" weight="light">
+              <Text size="small" weight="light">
                 <br />
                 par{' '}
-                <Text weight="bold" size="regular">
+                <Text weight="bold" size="medium">
                   Sick Magazine
                 </Text>
               </Text>
 
               <Text
                 tag="p"
-                size="tiny"
+                size="small"
                 className="k-u-margin-top-single k-u-margin-bottom-none"
               >
                 Édition Barkley.
               </Text>
               <List className="k-SubscriptionDetails__listItem">
-                <Text size="tiny">
+                <Text size="small">
                   Le livre Les Finisseurs (signé par les auteurs).
                 </Text>
-                <Text size="tiny">Le coffret édition limitée.</Text>
-                <Text size="tiny">
+                <Text size="small">Le coffret édition limitée.</Text>
+                <Text size="small">
                   Un tirage 20 x 30 cm (signé par Alexis Berg).
                 </Text>
-                <Text size="tiny">Le DVD du film Les Finisseurs</Text>
+                <Text size="small">Le DVD du film Les Finisseurs</Text>
               </List>
             </div>
           </div>
@@ -219,22 +219,22 @@ export const ContributorReward = () => {
           <Separator className="k-u-margin-top-singleHalf k-u-margin-bottom-singleHalf" />
           <div className="k-SubscriptionDetails__details">
             <div>
-              <Text size="tiny">Option</Text>
+              <Text size="small">Option</Text>
               <Text
                 tag="p"
                 weight="bold"
-                size="tiny"
+                size="small"
                 className="k-u-margin-none"
               >
                 Végétarien, bleu
               </Text>
             </div>
             <div>
-              <Text size="tiny">Mode de livraison</Text>
+              <Text size="small">Mode de livraison</Text>
               <Text
                 tag="p"
                 weight="bold"
-                size="tiny"
+                size="small"
                 className="k-u-margin-none"
               >
                 Livraison