+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - `card` HOC
   - `default-props` HOC
   - `trigger-enhancer` HOC
+- Change/Deprecation: `Typography`:
+  - Replace `tiny, regular, default, big` by `small, medium, large` prop size values.
 
 ## [9.6.0] - 2022-04-04